+++ CHANGELOG.md
@@ -19,6 +19,7 @@ dev-master
  - Traversable/Collection are now extendable.
  - Refactored ArrayExpression by creating ArrayItemExpression representing each element
  - Refactored OrderBy query segment by representing each function and direction as an OrderedFunction class
+ - Renamed Queries\Segments\Operation::getTraversable to getValues
  - 
 
 2.1.1 (22/5/14)